CORE-33719 added CustomerIdWrapper component
