+++ src/view/components/customerIdWrapper.styl
@@ -10,12 +10,4 @@ OF ANY KIND, either express or implied. See the License for the specific languag
 governing permissions and limitations under the License.
 */
 
-import fetch from "../../utils/fetch";
-
-export default () =>
-  fetch("/data/core/idnamespace/identities").then(response =>
-    response.map(namespace => ({
-      value: namespace.code,
-      label: namespace.code
-    }))
-  );
+@import "../global";
